refactor(pr): Updated render functions type and moved files around
