+++ app/scripts/modules/core/src/managed/constraints/ConstraintCard.less
